+++ src/utils/log.ts
@@ -8,7 +8,11 @@ export const error = (source: string, description: string) => console.error(`sou
 
 export const onDispatch = (ctx: Context, ev: EventData) => {
   console.log(`dispatched input '${ev[1]}' from space '${ev[0]}' with parameters: '${ev[2]}' and '${ev[3]}'`)
-  console.log(ctx.components[ev[0]].state)
+  if (ctx.components[ev[0]]) {
+    console.log(ctx.components[ev[0]].state)
+  } else {
+    console.log(`The space ${ev[0]} has removed`)
+  }
 }
 
 export const logFns = {